+++ lib/builder/selectors/sql_where_selector.dart
@@ -31,18 +31,13 @@ import 'package:flutter_pom/builder/selectors/sql_selector.dart';
 import 'package:flutter_pom/flutter_pom.dart';
 
 class SQLWhereSelector extends SQLSelector {
-
   SQLCondition condition;
 
   SQLWhereSelector(this.condition);
 
   String toSql() {
     var builder = <String>[];
-    builder.addAll([
-      SQLKeywords.where,
-      condition.toSql()
-    ]);
+    builder.addAll([SQLKeywords.where, condition.toSql()]);
     return builder.join(SQLTypes.separator);
   }
-
 }