+++ lib/model/database.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_pom/context/base_model_context.dart';
+import 'package:flutter_pom/context/migration_context.dart';
 import 'package:flutter_pom/context/model_context.dart';
 import 'package:flutter_pom/model/migration_info.dart';
 import 'package:flutter_pom/model/table.dart';
@@ -8,8 +9,8 @@ abstract class Database {
   Map<Type, Table> _tables;
   Map<Type, ModelContext> _modelTables = <Type, ModelContext>{};
 
-  //ModelContext<$MigrationInfo> _migrationContext;
-  //final $MigrationInfo _migrationInfo = $MigrationInfo();
+  ModelContext<$MigrationInfo> _migrationContext;
+  final $MigrationInfo _migrationInfo = $MigrationInfo();
 
   String _name;
 
@@ -24,8 +25,10 @@ abstract class Database {
   /// Gets or sets whether the connection gets opened automatically
   bool autoOpen = true;
 
+  bool enableMigration = false;
+
   /// Creates a new Instance
-  Database(String name, {this.autoOpen = true}) {
+  Database(String name, {this.autoOpen = true, this.enableMigration = false}) {
     _name = name;
 
     if (this.autoOpen) {
@@ -33,49 +36,51 @@ abstract class Database {
     }
   }
 
-  /*
   Future<void> _doMigrations() async {
     _migrationContext = ModelContext<$MigrationInfo>(this, _migrationInfo);
-    _migrationContext.create();
-
-    print(_tables.length);
-    print(_tables.entries.length);
+    await _migrationContext.create();
 
     _tables.forEach((t, v) async {
       var table = _tables[t];
 
       print("*** Running Migration for ${table.tableName} ***");
 
-      var tableInfo = await _migrationContext.getRange(
-          where: "${_migrationInfo.name.name} = '${table.tableName}'"
-      );
+      var tableInfo =
+          await _migrationContext.where((q) => q.name.value == table.tableName);
 
+      /// If there is exactly one entry inside the model cache table
       if (tableInfo.length == 1) {
+        /// Get item
         var migrationInfo = tableInfo.first;
 
+        /// if the table revision > than the stored revision
         if (migrationInfo.tableRevision.value < table.revision) {
-
+          /// run migration for every revision between current and target
           for (var migrateVersion = migrationInfo.tableRevision.value;
-          migrateVersion <= table.revision; migrateVersion++) {
-            await table.migrate(MigrationContext(this, table), migrationInfo.tableRevision.value, migrateVersion);
+              migrateVersion <= table.revision;
+              migrateVersion++) {
+            await table.migrate(MigrationContext(this, table),
+                migrationInfo.tableRevision.value, migrateVersion);
             migrationInfo.tableRevision.value = migrateVersion;
             await _migrationContext.update(migrationInfo);
           }
         } else if (migrationInfo.tableRevision.value > table.revision) {
-          throw AssertionError("The table metadata for '${table.tableName}' are ahead of table the actual revision");
+          throw AssertionError(
+              "The table metadata for '${table.tableName}' are ahead of table the actual revision");
         }
+        /// If there is no model cached we create a new entry
       } else if (tableInfo.length == 0) {
         var newMigrationInfo = $MigrationInfo();
         newMigrationInfo.name.value = table.tableName;
         newMigrationInfo.tableRevision.value = table.revision;
         await _migrationContext.put(newMigrationInfo);
-
+        /// if there is more than one model, this is no good!
       } else if (tableInfo.length > 1) {
+        print("There is more than one migration info. Removing all from database. Please restart the app.");
         await _migrationContext.deleteAll();
       }
     });
   }
-  */
 
   Future<void> _initializeDatabase() async {
     _tables = initializeDatabase();
@@ -84,7 +89,9 @@ abstract class Database {
       throw AssertionError("There was no table defined for the database.");
     }
 
-    //await _doMigrations();
+    if (enableMigration) {
+      await _doMigrations();
+    }
   }
 
   /// Initializes the database