+++ lib/model/table.dart
@@ -9,13 +9,14 @@ import 'package:flutter_pom/flutter_pom.dart';
 import 'package:flutter_pom/model/field.dart';
 
 abstract class Table {
-
   String _name;
+
   /// Returns the table name
   String get tableName => _name;
 
   /// Returns the fields
   List<Field> _fields;
+
   /// Returns the fields
   List<Field> get fields => _fields;
 
@@ -44,15 +45,15 @@ abstract class Table {
     }
   }
 
+  /// Returns whether there are duplicate fields defined
   bool get _hasDuplicateField {
-    for(var field in _fields) {
+    for (var field in _fields) {
       if (_fields.where((f) => f.name == field.name).length > 1) {
         return true;
       }
     }
     return false;
   }
-
 
   /// Returns if there is more than one primary key
   bool get _hasDuplicatePrimaryKey {
@@ -81,7 +82,8 @@ abstract class Table {
   int get revision;
 
   /// Migrates the table
-  Future<void> migrate(MigrationContext context, int fromRevision, int toRevision) async {
+  Future<void> migrate(
+      MigrationContext context, int fromRevision, int toRevision) async {
     var newFields = _fields.where((f) => f.revision == toRevision);
     if (newFields != null && newFields.length > 0) {
       for (var field in _fields) {
@@ -91,19 +93,22 @@ abstract class Table {
   }
 
   /* Static Members */
+  /// Maps values to an instance of a Dataset
   static Table map(Map<String, dynamic> data, Table table) {
     for (var key in data.keys) {
       var field = table._getField(key);
 
       if (field == null) {
-        throw MissingFieldError(table, 'Cannot map data to field "$key". Field was not found.');
+        throw MissingFieldError(
+            table, 'Cannot map data to field "$key". Field was not found.');
       }
 
       if (data[key] != null) {
         try {
           field.fromSqlCompatibleValue(data[key]);
         } catch (e) {
-          throw FieldConstraintError(field, 'Cannot map data to field "${field.name}". ${e.toString()}');
+          throw FieldConstraintError(field,
+              'Cannot map data to field "${field.name}". ${e.toString()}');
         }
       } else {
         field.fromSqlCompatibleValue(null);